+++ src/main/kotlin/jp/nephy/penicillin/core/emulation/Emulation.kt
@@ -1,7 +1,7 @@
 package jp.nephy.penicillin.core.emulation
 
 import io.ktor.http.Headers
-import jp.nephy.penicillin.core.Session
+import jp.nephy.penicillin.core.session.Session
 
 abstract class Emulation {
     abstract val session: Session
